refactor(adapters): isolate ZensicalAdapter using shared _utils functions to prevent regression in compat mode
