+++ src/zenzic/core/adapters/_zensical.py
@@ -40,8 +40,9 @@ from zenzic.core.adapters._base import BaseAdapter
 from zenzic.core.adapters._mkdocs_config import find_mkdocs_config_file, load_mkdocs_config
 
 
-from zenzic.core.adapters._utils import case_sensitive_exists, remap_to_default_locale
+from zenzic.core.adapters._utils import _extract_blog_dir, case_sensitive_exists, remap_to_default_locale
 from zenzic.core.exceptions import ZenzicConfigError
+
 from zenzic.models.config import BuildContext
 
 
@@ -431,12 +432,11 @@ class ZensicalAdapter(BaseAdapter):
     @property
     def dynamic_directories(self) -> set[Path]:
         """Return directories dynamically managed by plugins (e.g. blog posts)."""
-        from zenzic.core.adapters._mkdocs import _extract_blog_dir
-
         blog_prefix = _extract_blog_dir(self._doc_config)
         if blog_prefix:
             return {(self._docs_root / blog_prefix).resolve()}
         return set()
+
